Extract POST function in D1Detail && Resume needs POST function to do post to backend
